Moved query to query folder and removed hook file
