+++ jsapp/js/account/billingContextProvider.component.tsx
@@ -3,7 +3,7 @@ import {UsageContext, useUsage} from 'js/account/usage/useUsage.hook';
 import {ProductsContext, useProducts} from 'js/account/useProducts.hook';
 import {
   useOrganizationQuery,
-} from 'js/account/organizations/useOrganization.hook';
+} from 'js/query/queries/organizations.query';
 import sessionStore from 'js/stores/session';
 
 export const BillingContextProvider = (props: {children: ReactNode}) => {
